+++ TypeContractor.Tests/TypeScript/TypeScriptConverterTests.cs
@@ -33,13 +33,15 @@ public class TypeScriptConverterTests
         result.FullName.Should().Be("TypeContractor.Tests.TypeScript.TypeScriptConverterTests+SimpleTypes");
         result.IsEnum.Should().BeFalse();
         result.EnumMembers.Should().BeNull();
-        result.Properties.Should().HaveCount(3);
+        result.Properties.Should().HaveCount(5);
     }
 
     [Theory]
     [InlineData(0)]
     [InlineData(1)]
     [InlineData(2)]
+    [InlineData(3)]
+    [InlineData(4)]
     public void Converted_Simple_Properties_Looks_As_Expected(int propertyIndex)
     {
         var result = Sut.Convert(typeof(SimpleTypes));
@@ -80,6 +82,28 @@ public class TypeScriptConverterTests
                 prop.IsArray.Should().BeTrue();
                 prop.IsNullable.Should().BeFalse();
                 break;
+
+            case 3:
+                prop.SourceName.Should().Be("DoubleTime");
+                prop.SourceType.Should().Be(typeof(double));
+                prop.InnerSourceType.Should().BeNull();
+                prop.DestinationName.Should().Be("doubleTime");
+                prop.DestinationType.Should().Be("number");
+                prop.IsBuiltin.Should().BeTrue();
+                prop.IsArray.Should().BeFalse();
+                prop.IsNullable.Should().BeFalse();
+                break;
+
+            case 4:
+                prop.SourceName.Should().Be("TimeyWimeySpan");
+                prop.SourceType.Should().Be(typeof(TimeSpan));
+                prop.InnerSourceType.Should().BeNull();
+                prop.DestinationName.Should().Be("timeyWimeySpan");
+                prop.DestinationType.Should().Be("string");
+                prop.IsBuiltin.Should().BeTrue();
+                prop.IsArray.Should().BeFalse();
+                prop.IsNullable.Should().BeFalse();
+                break;
         }
     }
 
@@ -99,11 +123,13 @@ public class TypeScriptConverterTests
         var result = Sut.Convert(typeof(NestedInheritanceTest));
 
         result.Properties.Should().NotBeNull();
-        result.Properties.Should().HaveCount(5);
+        result.Properties.Should().HaveCount(7);
         result.Properties.Should()
             .Contain(x => x.SourceName == "StringProperty")
             .And.Contain(x => x.SourceName == "NumberProperty")
             .And.Contain(x => x.SourceName == "NumbersProperty")
+            .And.Contain(x => x.SourceName == "DoubleTime")
+            .And.Contain(x => x.SourceName == "TimeyWimeySpan")
             .And.Contain(x => x.SourceName == "InheritedProperty")
             .And.Contain(x => x.SourceName == "FinalProperty");
     }
@@ -191,6 +217,8 @@ public class TypeScriptConverterTests
         public string StringProperty { get; set; }
         public int? NumberProperty { get; set; }
         public IEnumerable<int> NumbersProperty { get; set; }
+        public double DoubleTime { get; set; }
+        public TimeSpan TimeyWimeySpan { get; set; }
     }
 
     private class TypeVisibility